Moved thread access checks to PHP.
Now all forumservices-queries go through a accesscheck before they are executed. Formely the thread access code could be bypassed by using the JS console but this should no longer be possible.

+++ DuggaSys/forum.js
@@ -44,6 +44,17 @@ function makeComment()
 	AJAXService("MAKECOMMENT",{threadId:querystring["threadId"],userID:userID,text:text},"MAKECOMMENT");
 }
 
+function accessDenied()
+{
+	var str = "<div class='err'>";
+			str +=	"<span style='font-weight:bold'>";
+			str +=		"Access denied!";
+			str	+=	"</span>";
+			str +=	array["accessDenied"];
+			str +="</div>";
+	$("#content").html(str);
+}
+
 
 //----------------------------------------
 // Renderer
@@ -52,45 +63,53 @@ function makeComment()
 function returnedThread(array)
 {
 	console.log(array);
-	$(".threadTopic").html(array["thread"]["topic"]);
-	$("#threadDescr").html(array["thread"]["description"]);
-	var str = "<span id='threadDate'>";
-			str += 	array["thread"]["datecreated"].substring(0, 16);
-			str += "</span> by <span id='threadCreator'>a97marbr</span>";
-	$("#threadDetails").html(str);
+	if (array["accessDenied"]){
+		accessDenied();
+	}else {
+		$(".threadTopic").html(array["thread"]["topic"]);
+		$("#threadDescr").html(array["thread"]["description"]);
+		var str = "<span id='threadDate'>";
+				str += 	array["thread"]["datecreated"].substring(0, 16);
+				str += "</span> by <span id='threadCreator'>a97marbr</span>";
+		$("#threadDetails").html(str);
+	}
 }
 
 function returnedComments(array)
 {
-	// Adds the comment header with the amount of comments.
-	var commentLength = array["comments"].length;
-	var threadCommentsHeaderStr = "<div id=\"threadCommentsHeader\"> Comments ("  +  commentLength  + ") </div>"
+	if (array["accessDenied"]){
+		accessDenied();
+	}else {
+		// Adds the comment header with the amount of comments.
+		var commentLength = array["comments"].length;
+		var threadCommentsHeaderStr = "<div id=\"threadCommentsHeader\"> Comments ("  +  commentLength  + ") </div>"
 
-	$("#threadComments").append(threadCommentsHeaderStr);
+		$("#threadComments").append(threadCommentsHeaderStr);
 
-	var threadCommentStr="";
-	threadCommentStr = "<div class=\"allComments\">";
+		var threadCommentStr="";
+		threadCommentStr = "<div class=\"allComments\">";
 
-	// Iterates through all the comments
-	$.each(array["comments"], function(index, value){
+		// Iterates through all the comments
+		$.each(array["comments"], function(index, value){
 
-		threadCommentStr +=
-		"<div class=\"threadComment\">" +
-			"<div class=\"commentDetails\"><span id=\"commentUser\">Skrivet av: " + value["uid"]  +   "</span></div>" +
-			"<div class=\"commentContent\"> <p>" +  value["text"]  + "</p></div>" +
-			"<div class=\"commentFooter\">" +
-				"<input class=\"submit-button\" type=\"button\" value=\"Reply\" onclick=\"replyUI();\">" +
-				"<input class=\"submit-button\" type=\"button\" value=\"Edit\" onclick=\"editUI();\">" +
-				"<input class=\"submit-button\" type=\"button\" value=\"Delete\" onclick=\"deleteComment();\">" +
-			"</div>" +
+			threadCommentStr +=
+			"<div class=\"threadComment\">" +
+				"<div class=\"commentDetails\"><span id=\"commentUser\">Skrivet av: " + value["uid"]  +   "</span></div>" +
+				"<div class=\"commentContent\"> <p>" +  value["text"]  + "</p></div>" +
+				"<div class=\"commentFooter\">" +
+					"<input class=\"submit-button\" type=\"button\" value=\"Reply\" onclick=\"replyUI();\">" +
+					"<input class=\"submit-button\" type=\"button\" value=\"Edit\" onclick=\"editUI();\">" +
+					"<input class=\"submit-button\" type=\"button\" value=\"Delete\" onclick=\"deleteComment();\">" +
+				"</div>" +
 
-			"<div class=\"commentDate\">" + (value["datecreated"]).substring(0,10) + "</div></div>";
+				"<div class=\"commentDate\">" + (value["datecreated"]).substring(0,10) + "</div></div>";
 
-		// Appends the comment
-		$("#threadComments").append(threadCommentStr);
-	});
+			// Appends the comment
+			$("#threadComments").append(threadCommentStr);
+		});
 
-	threadCommentStr += "</div>";
+		threadCommentStr += "</div>";
+	}
 }
 
 function showThread(thread)
@@ -104,14 +123,10 @@ function showComment(comment)
 	console.log(comment);
 }
 
-function error(errMsg)
+function error(xhr, status, error)
 {
-	console.log("wdawd");
-	var str = "<div class='err'>";
-			str +=	"<span style='font-weight:bold'>"
-			str +=		errMsg["errMsgBold"]
-			str	+=	"</span>";
-			str +=	errMsg["errMsgBody"];
-			str +="</div>";
-	$("#content").html(str);
+	console.log("ERROOR");
+	console.log(error);
+	console.log(status);
+	console.log(xhr);
 }
